requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
attrs==24.3.0
auth0-python==4.7.2
backoff==2.2.1
bcrypt==4.2.1
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.8
coloredlogs==15.0.1
crewai==0.98.0
cryptography==43.0.3
decorator==5.1.1
Deprecated==1.2.15
distro==1.9.0
docstring_parser==0.16
durationpy==0.9
et_xmlfile==2.0.0
executing==2.2.0
fastapi==0.115.6
filelock==3.17.0
filetype==1.2.0
flatbuffers==25.1.21
frozenlist==1.5.0
fsspec==2024.12.0
gast==0.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.78.0
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.14.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-generativeai==0.8.4
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.14.0
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.27.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
instructor==1.7.2
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
json_repair==0.35.0
jsonpatch==1.33
jsonpickle==4.0.1
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.8.0
kubernetes==31.0.0
langchain==0.3.15
langchain-core==0.3.31
langchain-google-genai==2.0.9
langchain-text-splitters==0.3.5
langsmith==0.3.1
libclang==18.1.1
litellm==1.57.4
load-dotenv==0.1.0
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
namex==0.0.8
networkx==3.4.2
numpy==2.0.2
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.60.0
openpyxl==3.1.5
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
opt_einsum==3.4.0
optree==0.14.0
orjson==3.10.15
overrides==7.7.0
packaging==24.2
parso==0.8.4
pdfminer.six==20231228
pdfplumber==0.11.5
pexpect==4.9.0
pillow==11.1.0
posthog==3.9.2
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.1
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytils==0.4.1
pyvis==0.3.2
PyYAML==6.0.2
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
setuptools==75.1.0
shapely==2.0.6
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
stack-data==0.6.3
starlette==0.41.3
sympy==1.13.3
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
tiktoken==0.8.0
tokenizers==0.21.0
tomli==2.2.1
tomli_w==1.2.0
tools==0.1.9
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uv==0.5.22
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
wcwidth==0.2.13
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
wheel==0.44.0
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
